fix signal_resample

+++ neurokit2/signal/signal_resample.py
@@ -128,7 +128,7 @@ def _resample_pandas(signal, desired_length):
     resampled_signal = pd.Series(signal, index=index)
 
     # Create resampling factor
-    resampling_factor = str(1/(desired_length / len(signal))) + "L"
+    resampling_factor = str(int(1/(desired_length / len(signal)))) + "L"
 
     # Resample
     resampled_signal = resampled_signal.resample(resampling_factor).bfill().values
